Remove extra code after local setup refactor
