--- chains/evm/calls/consts/feeHandler.go
@@ -1,13 +1,0 @@
-// The Licensed Work is (c) 2022 Sygma
-// SPDX-License-Identifier: BUSL-1.1
-
-package consts
-
-const FeeHandlerWithOracleABI = "[{\"inputs\":[{\"internalType\":\"address\",\"name\":\"bridgeAddress\",\"type\":\"address\"},{\"internalType\":\"address\",\"name\":\"feeHandlerRouterAddress\",\"type\":\"address\"}],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"}],\"name\":\"FeeCollected\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"recipient\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"amount\",\"type\":\"uint256\"}],\"name\":\"FeeDistributed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"previousAdminRole\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"newAdminRole\",\"type\":\"bytes32\"}],\"name\":\"RoleAdminChanged\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleGranted\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleRevoked\",\"type\":\"event\"},{\"inputs\":[],\"name\":\"DEFAULT_ADMIN_ROLE\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_bridgeAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_feeHandlerRouterAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_feePercent\",\"outputs\":[{\"internalType\":\"uint16\",\"name\":\"\",\"type\":\"uint16\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_gasUsed\",\"outputs\":[{\"internalType\":\"uint32\",\"name\":\"\",\"type\":\"uint32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_oracleAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"}],\"name\":\"getRoleAdmin\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"grantRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"hasRole\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"renounceRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"revokeRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes4\",\"name\":\"interfaceId\",\"type\":\"bytes4\"}],\"name\":\"supportsInterface\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"newAdmin\",\"type\":\"address\"}],\"name\":\"renounceAdmin\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"oracleAddress\",\"type\":\"address\"}],\"name\":\"setFeeOracle\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint32\",\"name\":\"gasUsed\",\"type\":\"uint32\"},{\"internalType\":\"uint16\",\"name\":\"feePercent\",\"type\":\"uint16\"}],\"name\":\"setFeeProperties\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"collectFee\",\"outputs\":[],\"stateMutability\":\"payable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"calculateFee\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"addrs\",\"type\":\"address[]\"},{\"internalType\":\"uint256[]\",\"name\":\"amounts\",\"type\":\"uint256[]\"}],\"name\":\"transferFee\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"}]"
-const FeeHandlerWithOracleBin = "0x60c06040523480156200001157600080fd5b5060405162001ee138038062001ee1833981016040819052620000349162000129565b6001600160a01b03808316608052811660a052620000546000336200005c565b505062000161565b6200006882826200006c565b5050565b6000828152602081815260408083206001600160a01b038516845290915290205460ff1662000068576000828152602081815260408083206001600160a01b03851684529091529020805460ff19166001179055620000c83390565b6001600160a01b0316816001600160a01b0316837f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d60405160405180910390a45050565b80516001600160a01b03811681146200012457600080fd5b919050565b600080604083850312156200013d57600080fd5b62000148836200010c565b915062000158602084016200010c565b90509250929050565b60805160a051611d3e620001a3600039600081816102a90152610a0e0152600081816101c801528181610732015281816109dc0152610cd90152611d3e6000f3fe6080604052600436106101095760003560e01c8063745e6b6111610095578063bff4275511610064578063bff4275514610320578063c297983f14610340578063d547741f14610360578063ef4f081f14610380578063fc818cfb146103bd57600080fd5b8063745e6b611461029757806391d14854146102cb578063a217fddf146102eb578063a8a989621461030057600080fd5b8063318c136e116100dc578063318c136e146101b657806336568abe146102025780635e1fab0f1461022257806369222948146102425780636fb7cb571461027757600080fd5b806301ffc9a71461010e578063248a9ca31461014357806325307065146101815780632f2ff15d14610196575b600080fd5b34801561011a57600080fd5b5061012e610129366004611741565b6103f6565b60405190151581526020015b60405180910390f35b34801561014f57600080fd5b5061017361015e36600461176b565b60009081526020819052604090206001015490565b60405190815260200161013a565b61019461018f3660046117ea565b61042d565b005b3480156101a257600080fd5b506101946101b136600461189b565b610519565b3480156101c257600080fd5b506101ea7f000000000000000000000000000000000000000000000000000000000000000081565b6040516001600160a01b03909116815260200161013a565b34801561020e57600080fd5b5061019461021d36600461189b565b610543565b34801561022e57600080fd5b5061019461023d3660046118cb565b6105c1565b34801561024e57600080fd5b5060015461026490600160c01b900461ffff1681565b60405161ffff909116815260200161013a565b34801561028357600080fd5b506001546101ea906001600160a01b031681565b3480156102a357600080fd5b506101ea7f000000000000000000000000000000000000000000000000000000000000000081565b3480156102d757600080fd5b5061012e6102e636600461189b565b610631565b3480156102f757600080fd5b50610173600081565b34801561030c57600080fd5b5061019461031b3660046118cb565b61065a565b34801561032c57600080fd5b5061019461033b36600461192d565b6106a3565b34801561034c57600080fd5b5061019461035b3660046119a7565b61091e565b34801561036c57600080fd5b5061019461037b36600461189b565b610988565b34801561038c57600080fd5b506103a061039b3660046117ea565b6109ad565b604080519283526001600160a01b0390911660208301520161013a565b3480156103c957600080fd5b506001546103e190600160a01b900463ffffffff1681565b60405163ffffffff909116815260200161013a565b60006001600160e01b03198216637965db0b60e01b148061042757506301ffc9a760e01b6001600160e01b03198316145b92915050565b6104356109d1565b34156104885760405162461bcd60e51b815260206004820152601a60248201527f636f6c6c6563744665653a206d73672e76616c756520213d203000000000000060448201526064015b60405180910390fd5b60008061049b8a8a8a8a8a8a8a8a610a93565b915091506104ab818b3085610e59565b604080516001600160a01b038c8116825260ff8c811660208401528b1682840152606082018a905260808201859052831660a082015290517fbd231b7fa4103e15e7a238c72f07e8aff310701af121895aa6c793b80245e4339181900360c00190a150505050505050505050565b60008281526020819052604090206001015461053481610e6d565b61053e8383610e7a565b505050565b6001600160a01b03811633146105b35760405162461bcd60e51b815260206004820152602f60248201527f416363657373436f6e74726f6c3a2063616e206f6e6c792072656e6f756e636560448201526e103937b632b9903337b91039b2b63360891b606482015260840161047f565b6105bd8282610efe565b5050565b336001600160a01b03821681141561061b5760405162461bcd60e51b815260206004820152601760248201527f43616e6e6f742072656e6f756e6365206f6e6573656c66000000000000000000604482015260640161047f565b610626600083610519565b6105bd600082610543565b6000918252602082815260408084206001600160a01b0393909316845291905290205460ff1690565b610665600033610631565b6106815760405162461bcd60e51b815260040161047f906119e5565b600180546001600160a01b0319166001600160a01b0392909216919091179055565b6106ae600033610631565b6106ca5760405162461bcd60e51b815260040161047f906119e5565b8281146107195760405162461bcd60e51b815260206004820152601f60248201527f61646472735b5d2c20616d6f756e74735b5d3a2064696666206c656e67746800604482015260640161047f565b6040516384db809f60e01b8152600481018690526000907f00000000000000000000000000000000000000000000000000000000000000006001600160a01b0316906384db809f90602401602060405180830381865afa158015610781573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906107a59190611a1c565b60405163014daabb60e31b8152600481018890529091506000906001600160a01b03831690630a6d55d890602401602060405180830381865afa1580156107f0573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906108149190611a1c565b905060005b858110156109145761086a8288888481811061083757610837611a39565b905060200201602081019061084c91906118cb565b87878581811061085e5761085e611a39565b90506020020135610f63565b7faaa40a232aaf133fdd28f3485f6fdd163514cfadbffa981f3610f42398efe34b8288888481811061089e5761089e611a39565b90506020020160208101906108b391906118cb565b8787858181106108c5576108c5611a39565b905060200201356040516108fa939291906001600160a01b039384168152919092166020820152604081019190915260600190565b60405180910390a18061090c81611a65565b915050610819565b5050505050505050565b610929600033610631565b6109455760405162461bcd60e51b815260040161047f906119e5565b6001805461ffff909216600160c01b0261ffff60c01b1963ffffffff909416600160a01b029390931665ffffffffffff60a01b1990921691909117919091179055565b6000828152602081905260409020600101546109a381610e6d565b61053e8383610efe565b6000806109c08a8a8a8a8a8a8a8a610a93565b915091509850989650505050505050565b336001600160a01b037f0000000000000000000000000000000000000000000000000000000000000000161480610a305750336001600160a01b037f000000000000000000000000000000000000000000000000000000000000000016145b610a915760405162461bcd60e51b815260206004820152602c60248201527f73656e646572206d75737420626520627269646765206f722066656520726f7560448201526b1d195c8818dbdb9d1c9858dd60a21b606482015260840161047f565b565b6000806101418314610ae75760405162461bcd60e51b815260206004820152601860248201527f496e636f72726563742066656544617461206c656e6774680000000000000000604482015260640161047f565b610b0b60405180606001604052806060815260200160608152602001600081525090565b610b1960e060008688611a80565b8080601f016020809104026020016040519081016040528093929190818152602001838380828437600092019190915250505090825250610b5f61012160e08688611a80565b8080601f0160208091040260200160405190810160405280939291908181526020018383808284376000920191909152505050506020820152610ba6846101218188611a80565b810190610bb3919061176b565b604082015280518051600091610bd191810160209081019101611ad0565b90508060600151421115610c1e5760405162461bcd60e51b81526020600482015260146024820152734f62736f6c657465206f7261636c65206461746160601b604482015260640161047f565b8a60ff16816080015160ff16148015610c4057508960ff168160a0015160ff16145b8015610c4f5750888160c00151145b610c9b5760405162461bcd60e51b815260206004820152601860248201527f496e636f7272656374206465706f73697420706172616d730000000000000000604482015260640161047f565b8151805160209182012090830151600154610cc09183916001600160a01b0316610f75565b6040516384db809f60e01b8152600481018b90526000907f00000000000000000000000000000000000000000000000000000000000000006001600160a01b0316906384db809f90602401602060405180830381865afa158015610d28573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610d4c9190611a1c565b60405163014daabb60e31b8152600481018d90529091506000906001600160a01b03831690630a6d55d890602401602060405180830381865afa158015610d97573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190610dbb9190611a1c565b90506000670de0b6b3a76400008560200151600160149054906101000a900463ffffffff1663ffffffff168760400151610df59190611b70565b610dff9190611b70565b610e099190611b8f565b600154604088015191925061271091610e2d91600160c01b900461ffff1690611b70565b610e379190611b8f565b975080881015610e45578097505b509450505050509850989650505050505050565b83610e6681858585610fd8565b5050505050565b610e778133611043565b50565b610e848282610631565b6105bd576000828152602081815260408083206001600160a01b03851684529091529020805460ff19166001179055610eba3390565b6001600160a01b0316816001600160a01b0316837f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d60405160405180910390a45050565b610f088282610631565b156105bd576000828152602081815260408083206001600160a01b0385168085529252808320805460ff1916905551339285917ff6391f5c32d9c69d2a47ea670b442974b53935d1edc7fd64eb21e047a839171b9190a45050565b82610f6f8184846110a7565b50505050565b6000610f8184846110d7565b9050816001600160a01b0316816001600160a01b031614610f6f5760405162461bcd60e51b8152602060048201526011602482015270496e76616c6964207369676e617475726560781b604482015260640161047f565b6040516001600160a01b0380851660248301528316604482015260648101829052610f6f9085906323b872dd60e01b906084015b60408051601f198184030181529190526020810180516001600160e01b03166001600160e01b0319909316929092179091526110fb565b61104d8282610631565b6105bd57611065816001600160a01b0316601461124d565b61107083602061124d565b604051602001611081929190611bdd565b60408051601f198184030181529082905262461bcd60e51b825261047f91600401611c52565b6040516001600160a01b03831660248201526044810182905261053e90849063a9059cbb60e01b9060640161100c565b60008060006110e685856113f0565b915091506110f381611460565b509392505050565b813b806111425760405162461bcd60e51b8152602060048201526015602482015274115490cc8c0e881b9bdd08184818dbdb9d1c9858dd605a1b604482015260640161047f565b600080846001600160a01b03168460405161115d9190611c85565b6000604051808303816000865af19150503d806000811461119a576040519150601f19603f3d011682016040523d82523d6000602084013e61119f565b606091505b5091509150816111e65760405162461bcd60e51b8152602060048201526012602482015271115490cc8c0e8818d85b1b0819985a5b195960721b604482015260640161047f565b805115610e6657808060200190518101906112019190611ca1565b610e665760405162461bcd60e51b815260206004820181905260248201527f45524332303a206f7065726174696f6e20646964206e6f742073756363656564604482015260640161047f565b6060600061125c836002611b70565b611267906002611cc3565b67ffffffffffffffff81111561127f5761127f611aaa565b6040519080825280601f01601f1916602001820160405280156112a9576020820181803683370190505b509050600360fc1b816000815181106112c4576112c4611a39565b60200101906001600160f81b031916908160001a905350600f60fb1b816001815181106112f3576112f3611a39565b60200101906001600160f81b031916908160001a9053506000611317846002611b70565b611322906001611cc3565b90505b600181111561139a576f181899199a1a9b1b9c1cb0b131b232b360811b85600f166010811061135657611356611a39565b1a60f81b82828151811061136c5761136c611a39565b60200101906001600160f81b031916908160001a90535060049490941c9361139381611cdb565b9050611325565b5083156113e95760405162461bcd60e51b815260206004820181905260248201527f537472696e67733a20686578206c656e67746820696e73756666696369656e74604482015260640161047f565b9392505050565b6000808251604114156114275760208301516040840151606085015160001a61141b8782858561161b565b94509450505050611459565b8251604014156114515760208301516040840151611446868383611708565b935093505050611459565b506000905060025b9250929050565b600081600481111561147457611474611cf2565b141561147d5750565b600181600481111561149157611491611cf2565b14156114df5760405162461bcd60e51b815260206004820152601860248201527f45434453413a20696e76616c6964207369676e61747572650000000000000000604482015260640161047f565b60028160048111156114f3576114f3611cf2565b14156115415760405162461bcd60e51b815260206004820152601f60248201527f45434453413a20696e76616c6964207369676e6174757265206c656e67746800604482015260640161047f565b600381600481111561155557611555611cf2565b14156115ae5760405162461bcd60e51b815260206004820152602260248201527f45434453413a20696e76616c6964207369676e6174757265202773272076616c604482015261756560f01b606482015260840161047f565b60048160048111156115c2576115c2611cf2565b1415610e775760405162461bcd60e51b815260206004820152602260248201527f45434453413a20696e76616c6964207369676e6174757265202776272076616c604482015261756560f01b606482015260840161047f565b6000807f7fffffffffffffffffffffffffffffff5d576e7357a4501ddfe92f46681b20a083111561165257506000905060036116ff565b8460ff16601b1415801561166a57508460ff16601c14155b1561167b57506000905060046116ff565b6040805160008082526020820180845289905260ff881692820192909252606081018690526080810185905260019060a0016020604051602081039080840390855afa1580156116cf573d6000803e3d6000fd5b5050604051601f1901519150506001600160a01b0381166116f8576000600192509250506116ff565b9150600090505b94509492505050565b6000806001600160ff1b0383168161172560ff86901c601b611cc3565b90506117338782888561161b565b935093505050935093915050565b60006020828403121561175357600080fd5b81356001600160e01b0319811681146113e957600080fd5b60006020828403121561177d57600080fd5b5035919050565b6001600160a01b0381168114610e7757600080fd5b60ff81168114610e7757600080fd5b60008083601f8401126117ba57600080fd5b50813567ffffffffffffffff8111156117d257600080fd5b60208301915083602082850101111561145957600080fd5b60008060008060008060008060c0898b03121561180657600080fd5b883561181181611784565b9750602089013561182181611799565b9650604089013561183181611799565b955060608901359450608089013567ffffffffffffffff8082111561185557600080fd5b6118618c838d016117a8565b909650945060a08b013591508082111561187a57600080fd5b506118878b828c016117a8565b999c989b5096995094979396929594505050565b600080604083850312156118ae57600080fd5b8235915060208301356118c081611784565b809150509250929050565b6000602082840312156118dd57600080fd5b81356113e981611784565b60008083601f8401126118fa57600080fd5b50813567ffffffffffffffff81111561191257600080fd5b6020830191508360208260051b850101111561145957600080fd5b60008060008060006060868803121561194557600080fd5b85359450602086013567ffffffffffffffff8082111561196457600080fd5b61197089838a016118e8565b9096509450604088013591508082111561198957600080fd5b50611996888289016118e8565b969995985093965092949392505050565b600080604083850312156119ba57600080fd5b823563ffffffff811681146119ce57600080fd5b9150602083013561ffff811681146118c057600080fd5b6020808252601e908201527f73656e64657220646f65736e277420686176652061646d696e20726f6c650000604082015260600190565b600060208284031215611a2e57600080fd5b81516113e981611784565b634e487b7160e01b600052603260045260246000fd5b634e487b7160e01b600052601160045260246000fd5b6000600019821415611a7957611a79611a4f565b5060010190565b60008085851115611a9057600080fd5b83861115611a9d57600080fd5b5050820193919092039150565b634e487b7160e01b600052604160045260246000fd5b8051611acb81611799565b919050565b600060e08284031215611ae257600080fd5b60405160e0810181811067ffffffffffffffff82111715611b1357634e487b7160e01b600052604160045260246000fd5b806040525082518152602083015160208201526040830151604082015260608301516060820152611b4660808401611ac0565b6080820152611b5760a08401611ac0565b60a082015260c083015160c08201528091505092915050565b6000816000190483118215151615611b8a57611b8a611a4f565b500290565b600082611bac57634e487b7160e01b600052601260045260246000fd5b500490565b60005b83811015611bcc578181015183820152602001611bb4565b83811115610f6f5750506000910152565b7f416363657373436f6e74726f6c3a206163636f756e7420000000000000000000815260008351611c15816017850160208801611bb1565b7001034b99036b4b9b9b4b733903937b6329607d1b6017918401918201528351611c46816028840160208801611bb1565b01602801949350505050565b6020815260008251806020840152611c71816040850160208701611bb1565b601f01601f19169190910160400192915050565b60008251611c97818460208701611bb1565b9190910192915050565b600060208284031215611cb357600080fd5b815180151581146113e957600080fd5b60008219821115611cd657611cd6611a4f565b500190565b600081611cea57611cea611a4f565b506000190190565b634e487b7160e01b600052602160045260246000fdfea26469706673582212200d157cbcce9cdb7ddf8057e6ded8b2cead186605c4feb3736ac7e31c8b32dfbc64736f6c634300080b0033"
-
-const BasicFeeHandlerABI = "[{\"inputs\":[{\"internalType\":\"address\",\"name\":\"bridgeAddress\",\"type\":\"address\"},{\"internalType\":\"address\",\"name\":\"feeHandlerRouterAddress\",\"type\":\"address\"}],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"newFee\",\"type\":\"uint256\"}],\"name\":\"FeeChanged\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"}],\"name\":\"FeeCollected\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"recipient\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"amount\",\"type\":\"uint256\"}],\"name\":\"FeeDistributed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleGranted\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleRevoked\",\"type\":\"event\"},{\"inputs\":[],\"name\":\"DEFAULT_ADMIN_ROLE\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_bridgeAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_fee\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_feeHandlerRouterAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"}],\"name\":\"getRoleAdmin\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"index\",\"type\":\"uint256\"}],\"name\":\"getRoleMember\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"}],\"name\":\"getRoleMemberCount\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"getRoleMemberIndex\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"grantRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"hasRole\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"renounceRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"revokeRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"newAdmin\",\"type\":\"address\"}],\"name\":\"renounceAdmin\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"collectFee\",\"outputs\":[],\"stateMutability\":\"payable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"calculateFee\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"newFee\",\"type\":\"uint256\"}],\"name\":\"changeFee\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address payable[]\",\"name\":\"addrs\",\"type\":\"address[]\"},{\"internalType\":\"uint256[]\",\"name\":\"amounts\",\"type\":\"uint256[]\"}],\"name\":\"transferFee\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"}]"
-const BasicFeeHandlerBin = "0x60c06040523480156200001157600080fd5b506040516200121738038062001217833981016040819052620000349162000162565b6001600160a01b03808316608052811660a052620000546000336200005c565b50506200019a565b6200006882826200006c565b5050565b6000828152602081815260409091206200009191839062000a00620000d3821b17901c565b15620000685760405133906001600160a01b0383169084907f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d90600090a45050565b6000620000ea836001600160a01b038416620000f3565b90505b92915050565b60008181526001830160205260408120546200013c57508154600181810184556000848152602080822090930184905584548482528286019093526040902091909155620000ed565b506000620000ed565b80516001600160a01b03811681146200015d57600080fd5b919050565b600080604083850312156200017657600080fd5b620001818362000145565b9150620001916020840162000145565b90509250929050565b60805160a051611049620001ce600039600081816102590152610a5201526000818161018d0152610a2001526110496000f3fe6080604052600436106100fe5760003560e01c8063745e6b6111610095578063a217fddf11610064578063a217fddf146102eb578063c5b37c2214610300578063ca15c87314610316578063d547741f14610336578063ef4f081f1461035657600080fd5b8063745e6b61146102475780639010d07c1461027b578063904e9b2b1461029b57806391d14854146102bb57600080fd5b806336568abe116100d157806336568abe146101c75780634e0df3f6146101e75780635e1fab0f146102075780636a1db1bf1461022757600080fd5b8063248a9ca31461010357806325307065146101465780632f2ff15d1461015b578063318c136e1461017b575b600080fd5b34801561010f57600080fd5b5061013361011e366004610d42565b60009081526020819052604090206002015490565b6040519081526020015b60405180910390f35b610159610154366004610dd2565b6103a0565b005b34801561016757600080fd5b50610159610176366004610e7f565b610465565b34801561018757600080fd5b506101af7f000000000000000000000000000000000000000000000000000000000000000081565b6040516001600160a01b03909116815260200161013d565b3480156101d357600080fd5b506101596101e2366004610e7f565b6104f3565b3480156101f357600080fd5b50610133610202366004610e7f565b61056d565b34801561021357600080fd5b50610159610222366004610eaf565b610598565b34801561023357600080fd5b50610159610242366004610d42565b610608565b34801561025357600080fd5b506101af7f000000000000000000000000000000000000000000000000000000000000000081565b34801561028757600080fd5b506101af610296366004610ecc565b6106ec565b3480156102a757600080fd5b506101596102b6366004610f33565b61070b565b3480156102c757600080fd5b506102db6102d6366004610e7f565b610950565b604051901515815260200161013d565b3480156102f757600080fd5b50610133600081565b34801561030c57600080fd5b5061013360015481565b34801561032257600080fd5b50610133610331366004610d42565b610968565b34801561034257600080fd5b50610159610351366004610e7f565b61097f565b34801561036257600080fd5b50610383610371366004610dd2565b60015460009850989650505050505050565b604080519283526001600160a01b0390911660208301520161013d565b6103a8610a15565b60015434146103f75760405162461bcd60e51b8152602060048201526016602482015275125b98dbdc9c9958dd08199959481cdd5c1c1b1a595960521b60448201526064015b60405180910390fd5b600154604080516001600160a01b038b16815260ff8a81166020830152891681830152606081018890526080810192909252600060a0830152517fbd231b7fa4103e15e7a238c72f07e8aff310701af121895aa6c793b80245e4339181900360c00190a15050505050505050565b6000828152602081905260409020600201546104819033610950565b6104e55760405162461bcd60e51b815260206004820152602f60248201527f416363657373436f6e74726f6c3a2073656e646572206d75737420626520616e60448201526e0818591b5a5b881d1bc819dc985b9d608a1b60648201526084016103ee565b6104ef8282610ad7565b5050565b6001600160a01b03811633146105635760405162461bcd60e51b815260206004820152602f60248201527f416363657373436f6e74726f6c3a2063616e206f6e6c792072656e6f756e636560448201526e103937b632b9903337b91039b2b63360891b60648201526084016103ee565b6104ef8282610b30565b6000828152602081815260408083206001600160a01b03851684526001019091529020545b92915050565b336001600160a01b0382168114156105f25760405162461bcd60e51b815260206004820152601760248201527f43616e6e6f742072656e6f756e6365206f6e6573656c6600000000000000000060448201526064016103ee565b6105fd600083610465565b6104ef6000826104f3565b610613600033610950565b61065f5760405162461bcd60e51b815260206004820152601e60248201527f73656e64657220646f65736e277420686176652061646d696e20726f6c65000060448201526064016103ee565b8060015414156106b15760405162461bcd60e51b815260206004820152601f60248201527f43757272656e742066656520697320657175616c20746f206e6577206665650060448201526064016103ee565b60018190556040518181527f6bbc57480a46553fa4d156ce702beef5f3ad66303b0ed1a5d4cb44966c6584c39060200160405180910390a150565b60008281526020819052604081206107049083610b89565b9392505050565b610716600033610950565b6107625760405162461bcd60e51b815260206004820152601e60248201527f73656e64657220646f65736e277420686176652061646d696e20726f6c65000060448201526064016103ee565b8281146107b15760405162461bcd60e51b815260206004820152601f60248201527f61646472735b5d2c20616d6f756e74735b5d3a2064696666206c656e6774680060448201526064016103ee565b60005b838110156109495760008585838181106107d0576107d0610f9f565b90506020020160208101906107e59190610eaf565b6001600160a01b031684848481811061080057610800610f9f565b9050602002013560405160006040518083038185875af1925050503d8060008114610847576040519150601f19603f3d011682016040523d82523d6000602084013e61084c565b606091505b505090508061089d5760405162461bcd60e51b815260206004820152601960248201527f466565206574686572207472616e73666572206661696c65640000000000000060448201526064016103ee565b7faaa40a232aaf133fdd28f3485f6fdd163514cfadbffa981f3610f42398efe34b60008787858181106108d2576108d2610f9f565b90506020020160208101906108e79190610eaf565b8686868181106108f9576108f9610f9f565b9050602002013560405161092e939291906001600160a01b039384168152919092166020820152604081019190915260600190565b60405180910390a1508061094181610fcb565b9150506107b4565b5050505050565b60008281526020819052604081206107049083610b95565b600081815260208190526040812061059290610bb7565b60008281526020819052604090206002015461099b9033610950565b6105635760405162461bcd60e51b815260206004820152603060248201527f416363657373436f6e74726f6c3a2073656e646572206d75737420626520616e60448201526f2061646d696e20746f207265766f6b6560801b60648201526084016103ee565b6000610704836001600160a01b038416610bc1565b336001600160a01b037f0000000000000000000000000000000000000000000000000000000000000000161480610a745750336001600160a01b037f000000000000000000000000000000000000000000000000000000000000000016145b610ad55760405162461bcd60e51b815260206004820152602c60248201527f73656e646572206d75737420626520627269646765206f722066656520726f7560448201526b1d195c8818dbdb9d1c9858dd60a21b60648201526084016103ee565b565b6000828152602081905260409020610aef9082610a00565b156104ef5760405133906001600160a01b0383169084907f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d90600090a45050565b6000828152602081905260409020610b489082610c10565b156104ef5760405133906001600160a01b0383169084907ff6391f5c32d9c69d2a47ea670b442974b53935d1edc7fd64eb21e047a839171b90600090a45050565b60006107048383610c25565b6001600160a01b03811660009081526001830160205260408120541515610704565b6000610592825490565b6000818152600183016020526040812054610c0857508154600181810184556000848152602080822090930184905584548482528286019093526040902091909155610592565b506000610592565b6000610704836001600160a01b038416610c4f565b6000826000018281548110610c3c57610c3c610f9f565b9060005260206000200154905092915050565b60008181526001830160205260408120548015610d38576000610c73600183610fe6565b8554909150600090610c8790600190610fe6565b9050818114610cec576000866000018281548110610ca757610ca7610f9f565b9060005260206000200154905080876000018481548110610cca57610cca610f9f565b6000918252602080832090910192909255918252600188019052604090208390555b8554869080610cfd57610cfd610ffd565b600190038181906000526020600020016000905590558560010160008681526020019081526020016000206000905560019350505050610592565b6000915050610592565b600060208284031215610d5457600080fd5b5035919050565b6001600160a01b0381168114610d7057600080fd5b50565b803560ff81168114610d8457600080fd5b919050565b60008083601f840112610d9b57600080fd5b50813567ffffffffffffffff811115610db357600080fd5b602083019150836020828501011115610dcb57600080fd5b9250929050565b60008060008060008060008060c0898b031215610dee57600080fd5b8835610df981610d5b565b9750610e0760208a01610d73565b9650610e1560408a01610d73565b955060608901359450608089013567ffffffffffffffff80821115610e3957600080fd5b610e458c838d01610d89565b909650945060a08b0135915080821115610e5e57600080fd5b50610e6b8b828c01610d89565b999c989b5096995094979396929594505050565b60008060408385031215610e9257600080fd5b823591506020830135610ea481610d5b565b809150509250929050565b600060208284031215610ec157600080fd5b813561070481610d5b565b60008060408385031215610edf57600080fd5b50508035926020909101359150565b60008083601f840112610f0057600080fd5b50813567ffffffffffffffff811115610f1857600080fd5b6020830191508360208260051b8501011115610dcb57600080fd5b60008060008060408587031215610f4957600080fd5b843567ffffffffffffffff80821115610f6157600080fd5b610f6d88838901610eee565b90965094506020870135915080821115610f8657600080fd5b50610f9387828801610eee565b95989497509550505050565b634e487b7160e01b600052603260045260246000fd5b634e487b7160e01b600052601160045260246000fd5b6000600019821415610fdf57610fdf610fb5565b5060010190565b600082821015610ff857610ff8610fb5565b500390565b634e487b7160e01b600052603160045260246000fdfea2646970667358221220fb28b343ba429157b3fa1665842837f49a366a85f88238af3f0ccd0528cce55e64736f6c634300080b0033"
-
-const FeeRouterABI = "[{\"inputs\":[{\"internalType\":\"address\",\"name\":\"bridgeAddress\",\"type\":\"address\"}],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"newFee\",\"type\":\"uint256\"}],\"name\":\"FeeChanged\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"indexed\":false,\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"}],\"name\":\"FeeCollected\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"recipient\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"amount\",\"type\":\"uint256\"}],\"name\":\"FeeDistributed\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleGranted\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"},{\"indexed\":true,\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"}],\"name\":\"RoleRevoked\",\"type\":\"event\"},{\"inputs\":[],\"name\":\"DEFAULT_ADMIN_ROLE\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[],\"name\":\"_bridgeAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint8\",\"name\":\"\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"name\":\"_domainResourceIDToFeeHandlerAddress\",\"outputs\":[{\"internalType\":\"contract IFeeHandler\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"}],\"name\":\"getRoleAdmin\",\"outputs\":[{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"uint256\",\"name\":\"index\",\"type\":\"uint256\"}],\"name\":\"getRoleMember\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"}],\"name\":\"getRoleMemberCount\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"getRoleMemberIndex\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"\",\"type\":\"uint256\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"grantRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"hasRole\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"renounceRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"bytes32\",\"name\":\"role\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"account\",\"type\":\"address\"}],\"name\":\"revokeRole\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"contract IFeeHandler\",\"name\":\"handlerAddress\",\"type\":\"address\"}],\"name\":\"adminSetResourceHandler\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"collectFee\",\"outputs\":[],\"stateMutability\":\"payable\",\"type\":\"function\"},{\"inputs\":[{\"internalType\":\"address\",\"name\":\"sender\",\"type\":\"address\"},{\"internalType\":\"uint8\",\"name\":\"fromDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint8\",\"name\":\"destinationDomainID\",\"type\":\"uint8\"},{\"internalType\":\"bytes32\",\"name\":\"resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes\",\"name\":\"depositData\",\"type\":\"bytes\"},{\"internalType\":\"bytes\",\"name\":\"feeData\",\"type\":\"bytes\"}],\"name\":\"calculateFee\",\"outputs\":[{\"internalType\":\"uint256\",\"name\":\"fee\",\"type\":\"uint256\"},{\"internalType\":\"address\",\"name\":\"tokenAddress\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]"
-const FeeRouterBin = "0x60a06040523480156200001157600080fd5b5060405162000ed338038062000ed383398101604081905262000034916200013e565b6001600160a01b0381166080526200004e60003362000055565b5062000170565b62000061828262000065565b5050565b6000828152602081815260409091206200008a918390620006c5620000cc821b17901c565b15620000615760405133906001600160a01b0383169084907f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d90600090a45050565b6000620000e3836001600160a01b038416620000ec565b90505b92915050565b60008181526001830160205260408120546200013557508154600181810184556000848152602080822090930184905584548482528286019093526040902091909155620000e6565b506000620000e6565b6000602082840312156200015157600080fd5b81516001600160a01b03811681146200016957600080fd5b9392505050565b608051610d40620001936000396000818161018c015261073e0152610d406000f3fe6080604052600436106100dd5760003560e01c80634e0df3f61161007f578063a217fddf11610059578063a217fddf14610297578063ca15c873146102ac578063d547741f146102cc578063ef4f081f146102ec57600080fd5b80634e0df3f6146102275780639010d07c1461024757806391d148541461026757600080fd5b80632f2ff15d116100bb5780632f2ff15d1461015a578063318c136e1461017a57806336568abe146101c65780633d94ebc6146101e657600080fd5b80631af5fe38146100e2578063248a9ca3146101045780632530706514610147575b600080fd5b3480156100ee57600080fd5b506101026100fd366004610a44565b610329565b005b34801561011057600080fd5b5061013461011f366004610a84565b60009081526020819052604090206002015490565b6040519081526020015b60405180910390f35b610102610155366004610ae6565b61036e565b34801561016657600080fd5b50610102610175366004610b93565b610411565b34801561018657600080fd5b506101ae7f000000000000000000000000000000000000000000000000000000000000000081565b6040516001600160a01b03909116815260200161013e565b3480156101d257600080fd5b506101026101e1366004610b93565b6104a4565b3480156101f257600080fd5b506101ae610201366004610bc3565b60016020908152600092835260408084209091529082529020546001600160a01b031681565b34801561023357600080fd5b50610134610242366004610b93565b61051e565b34801561025357600080fd5b506101ae610262366004610bed565b610549565b34801561027357600080fd5b50610287610282366004610b93565b610568565b604051901515815260200161013e565b3480156102a357600080fd5b50610134600081565b3480156102b857600080fd5b506101346102c7366004610a84565b610580565b3480156102d857600080fd5b506101026102e7366004610b93565b610597565b3480156102f857600080fd5b5061030c610307366004610ae6565b610618565b604080519283526001600160a01b0390911660208301520161013e565b6103316106da565b60ff9290921660009081526001602090815260408083209383529290522080546001600160a01b0319166001600160a01b03909216919091179055565b610376610733565b60ff8616600090815260016020908152604080832088845290915290819020549051632530706560e01b81526001600160a01b0390911690819063253070659034906103d4908d908d908d908d908d908d908d908d90600401610c38565b6000604051808303818588803b1580156103ed57600080fd5b505af1158015610401573d6000803e3d6000fd5b5050505050505050505050505050565b60008281526020819052604090206002015461042d9033610568565b6104965760405162461bcd60e51b815260206004820152602f60248201527f416363657373436f6e74726f6c3a2073656e646572206d75737420626520616e60448201526e0818591b5a5b881d1bc819dc985b9d608a1b60648201526084015b60405180910390fd5b6104a082826107ab565b5050565b6001600160a01b03811633146105145760405162461bcd60e51b815260206004820152602f60248201527f416363657373436f6e74726f6c3a2063616e206f6e6c792072656e6f756e636560448201526e103937b632b9903337b91039b2b63360891b606482015260840161048d565b6104a08282610804565b6000828152602081815260408083206001600160a01b03851684526001019091529020545b92915050565b6000828152602081905260408120610561908361085d565b9392505050565b60008281526020819052604081206105619083610869565b60008181526020819052604081206105439061088b565b6000828152602081905260409020600201546105b39033610568565b6105145760405162461bcd60e51b815260206004820152603060248201527f416363657373436f6e74726f6c3a2073656e646572206d75737420626520616e60448201526f2061646d696e20746f207265766f6b6560801b606482015260840161048d565b60ff8616600090815260016020908152604080832088845290915280822054905163ef4f081f60e01b815282916001600160a01b031690819063ef4f081f90610673908e908e908e908e908e908e908e908e90600401610c38565b6040805180830381865afa15801561068f573d6000803e3d6000fd5b505050506040513d601f19601f820116820180604052508101906106b39190610c94565b92509250509850989650505050505050565b6000610561836001600160a01b038416610895565b6106e5600033610568565b6107315760405162461bcd60e51b815260206004820152601e60248201527f73656e64657220646f65736e277420686176652061646d696e20726f6c650000604482015260640161048d565b565b336001600160a01b037f000000000000000000000000000000000000000000000000000000000000000016146107315760405162461bcd60e51b815260206004820152601e60248201527f73656e646572206d7573742062652062726964676520636f6e74726163740000604482015260640161048d565b60008281526020819052604090206107c390826106c5565b156104a05760405133906001600160a01b0383169084907f2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d90600090a45050565b600082815260208190526040902061081c90826108e4565b156104a05760405133906001600160a01b0383169084907ff6391f5c32d9c69d2a47ea670b442974b53935d1edc7fd64eb21e047a839171b90600090a45050565b600061056183836108f9565b6001600160a01b03811660009081526001830160205260408120541515610561565b6000610543825490565b60008181526001830160205260408120546108dc57508154600181810184556000848152602080822090930184905584548482528286019093526040902091909155610543565b506000610543565b6000610561836001600160a01b038416610923565b600082600001828154811061091057610910610cb9565b9060005260206000200154905092915050565b60008181526001830160205260408120548015610a0c576000610947600183610ccf565b855490915060009061095b90600190610ccf565b90508181146109c057600086600001828154811061097b5761097b610cb9565b906000526020600020015490508087600001848154811061099e5761099e610cb9565b6000918252602080832090910192909255918252600188019052604090208390555b85548690806109d1576109d1610cf4565b600190038181906000526020600020016000905590558560010160008681526020019081526020016000206000905560019350505050610543565b6000915050610543565b803560ff81168114610a2757600080fd5b919050565b6001600160a01b0381168114610a4157600080fd5b50565b600080600060608486031215610a5957600080fd5b610a6284610a16565b9250602084013591506040840135610a7981610a2c565b809150509250925092565b600060208284031215610a9657600080fd5b5035919050565b60008083601f840112610aaf57600080fd5b50813567ffffffffffffffff811115610ac757600080fd5b602083019150836020828501011115610adf57600080fd5b9250929050565b60008060008060008060008060c0898b031215610b0257600080fd5b8835610b0d81610a2c565b9750610b1b60208a01610a16565b9650610b2960408a01610a16565b955060608901359450608089013567ffffffffffffffff80821115610b4d57600080fd5b610b598c838d01610a9d565b909650945060a08b0135915080821115610b7257600080fd5b50610b7f8b828c01610a9d565b999c989b5096995094979396929594505050565b60008060408385031215610ba657600080fd5b823591506020830135610bb881610a2c565b809150509250929050565b60008060408385031215610bd657600080fd5b610bdf83610a16565b946020939093013593505050565b60008060408385031215610c0057600080fd5b50508035926020909101359150565b81835281816020850137506000828201602090810191909152601f909101601f19169091010190565b60018060a01b038916815260ff8816602082015260ff8716604082015285606082015260c060808201526000610c7260c083018688610c0f565b82810360a0840152610c85818587610c0f565b9b9a5050505050505050505050565b60008060408385031215610ca757600080fd5b825191506020830151610bb881610a2c565b634e487b7160e01b600052603260045260246000fd5b600082821015610cef57634e487b7160e01b600052601160045260246000fd5b500390565b634e487b7160e01b600052603160045260246000fdfea2646970667358221220f0b560115748f4f59a1033f1c194fecd95aadc73361f837e9f46d1ab221450a964736f6c634300080b0033"
