+++ example/cmd/cmd.go
@@ -4,11 +4,9 @@
 package cmd
 
 import (
-	deploy2 "github.com/ChainSafe/sygma-relayer/chains/evm/cli/deploy"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/deploy"
 	"github.com/ChainSafe/chainbridge-core/flags"
 
 	"github.com/ChainSafe/sygma-relayer/example/app"
@@ -36,7 +34,7 @@ func init() {
 }
 
 func Execute() {
-	rootCMD.AddCommand(runCMD, deploy.DeployEVM, deploy2.LocalSetupCmd)
+	rootCMD.AddCommand(runCMD)
 	if err := rootCMD.Execute(); err != nil {
 		log.Fatal().Err(err).Msg("failed to execute root cmd")
 	}
